create_img.sh: add more packages

+++ create_img.sh
@@ -16,15 +16,17 @@ function finish {
 trap finish EXIT
 
 # Creat the image
-qemu-img create $IMG 1g
-mkfs.ext2 $IMG
+if [ ! -e $IMG ]; then
+	qemu-img create $IMG 1g
+	mkfs.ext2 $IMG
+fi
 
 # Mount it
 MOUNTED=1
 sudo mount -o loop $IMG $DIR
 
 # Create the contens
-sudo debootstrap --arch amd64 stretch $DIR
+sudo debootstrap --arch amd64 --include trace-cmd,kernelshark stretch $DIR
 
 # Setup root user
 sudo chroot $DIR
